apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
boto3==1.34.156
redis==5.0.8
rq==1.16.2
pgvector==0.2.5
openai==1.42.0
google-generativeai==0.8.3
python-multipart==0.0.9
pydantic==2.8.2
email-validator==2.2.0
pytest==8.3.2
httpx==0.27.2
websockets==12.0
